Fix: change dtype of node attrs for cueq-compatible compilation

+++ src/mlcg/nn/mace.py
@@ -899,7 +899,7 @@ class RBFRegularizedMACE(StandardMACE):
             zip(self.interactions, self.products, self.readouts)
         ):
             node_feats, sc = interaction(
-                node_attrs=node_attrs,
+                node_attrs=node_attrs.to(dtype=node_feats.dtype),
                 node_feats=node_feats,
                 edge_attrs=edge_attrs,
                 edge_feats=edge_feats[i],
